requirements.txt
streamlit
pandas
rdkit-pypi

